refactor: refactor data export and import
