+++ src/app/catalog/containers/categories/categories.component.scss
@@ -1,4 +1,4 @@
-@import '../../../shared/styles/variables.scss';
+@import '/src/app/shared/styles/variables';
 
 :host {
   display: flex;
