+++ src/app/catalog/containers/category-item/category-item.component.scss
@@ -1,4 +1,4 @@
-@import '../../../shared/styles/variables.scss';
+@import '/src/app/shared/styles/variables';
 
 :host {
   flex: 1 1 auto;
