+++ src/app/catalog/containers/category-item/category-item.component.ts
@@ -1,5 +1,5 @@
 import { Component, ChangeDetectionStrategy, Input } from '@angular/core';
-import { Category } from '@app/catalog/models/categories';
+import { Category } from '@app/catalog/models';
 
 @Component({
   selector: 'lc-category-item',
