+++ src/app/catalog/models/index.ts
@@ -1,1 +1,2 @@
 export * from './state.model';
+export * from './categories';
