+++ src/app/catalog/store/index.ts
@@ -1,2 +1,4 @@
 export * from './state';
 export * from './actions';
+export * from './state-categories';
+export * from './actions-categories';
